Added documentation, quantity parsing functions

+++ application/controllers/Buy_Transaction_Controller.php
@@ -6,7 +6,13 @@ class Buy_Transaction_Controller extends REST_Controller {
 
    /**
     * This method receives a JSON from the Android App containing Buy Transactions
+    * Sample JSON (As of 11/1):
+    * [{"buy_transaction_ts":"2015-10-31 19:22:05","id_number":133821,"shop_terminal_id":"001"}]
     *
+    * TODO:
+    *  Fix Buy_Transaction_ID Formatting (2->5)
+    *
+    * @flow: App->Server
     * @post_params: none
     * @output: Outputs decoded JSON array to Model
     */
@@ -25,7 +31,7 @@ class Buy_Transaction_Controller extends REST_Controller {
         foreach($parsable as $item) {
             $buy_transaction_ts = substr ($item, 23, 19);
             $id_number = substr ($item, 56, 6);
-            $shop_terminal_id = substr ($item, 83, 3);
+            $shop_terminal_id = (int)substr ($item, 83, 3);
             $this->Buy_Transaction_Model->sync_buy_transactions($buy_transaction_ts, $id_number, $shop_terminal_id);
         }
